requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
supabase>=2.0.0
reportlab>=4.0.0
requests>=2.28.0
Pillow>=10.0.0
streamlit-drawable-canvas>=0.9.3
python-docx>=1.0.0
PyPDF2>=3.0.0
